Move models from History api entry to models dedicated package
